pheap: finally fix
some assumptions were wrong, others were unverified.  three pointers
are used to represent the graph: parent, left (children), right (older
sibling).  siblings are stored as a linked list, with the most recent
one being in front.  most importantly, parent is either the previous
sibling in the list, or if it's the first node, the parent in the
tree.  essentially, this is a doubly linked list, with the parent
pointer helping efficiency despite the overload.  the biggest source
of issue was managing links correctly.

mergequeue takes the roots of two heaps.  a root may not by definition
have a parent or sibling (right pointer), but the left trees must be
taken care of.  mergepairs takes the *root* of a heap (when popping,
the root node is slashed off), splits, then merges recursively from
left to right.  here right nodes must be handled correctly to avoid
dropping entire branches of the tree and maintain parent links
correctly.  split trees are cut off from each other; the first two
heaps are merged together, and a third is recursively split and merged
the same way.  finally, decreasekey must preserve and fix the siblings
list, cut off the node, then merge the new heap with the root.
description subject to potential bugs, ymmv.

all relatively simple issues...  interesting future directions: smooth
heaps (2018 Kozma and Saranurak), related slim heaps.