add dynar: generic dynamic arrays using macros
combining ideas from klib and stb; having extra memory for a header
prepended to the vector is a good idea used elsewhere (pool, zalloc
here, etc.) which together with sizeof simplifies a lot the api, which
is in turn the biggest problem with klib's kvec.  not psyched about
using macros, but that's the way the cookie crumbles.